Rename Profile screen to Account, add "View my profile" link
The old Profile screen mixed settings (name editing, theme, notifications)
with identity. Now Account owns settings and links to the public player
profile via "View my profile". Route changed from /profile to /account.
